+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+### 1.1.2
+- `ActionCheckPieceCount` - allows for win conditions based on elimination of arbitrary pieces.
+- Variant definition: Kinglet (`MiscVariants.kinglet()`) - game is won when opponent has no pawns.
+
 ### 1.1.1
 - Custom drop move generation is now possible with `Variant.handOptions.dropBuilder`.
 - Added variant definitions: Mini Xiangqi (`Xiangqi.mini()`), Manchu (`Xiangqi.manchu()`), Hoppel-Poppel (`MiscVariants.hoppelPoppel()`), Shako (`LargeVariants.shako()`), Dobutsu (`Dobutsu.dobutsu()`).
 - Fixed a bug in serialisation of pieces with limited promo options.
-- Experimental (incomplete) Shogi and support.
+- Experimental (incomplete) Shogi support.
 - `Variant.withCampMate()` - helper method to add the campmate end condition to a variant.
 - `Variant.withPieces()` and `Variant.withPiecesRemoved()` helpers.
 - The play example can now load JSON variants.